requirements.txt
PySimpleGUI
PyMuPDF
Pillow
svg_stack